+++ Specs/Scene/TileBoundingBoxSpec.js
@@ -8,7 +8,7 @@ defineSuite([
         'Core/Math',
         'Core/Rectangle',
         'Scene/SceneMode',
-        'Specs/createFrameState'
+        'Specs/createScene'
     ], function(
         TileBoundingBox,
         Cartesian2,
@@ -18,15 +18,17 @@ defineSuite([
         CesiumMath,
         Rectangle,
         SceneMode,
-        createFrameState) {
+        createScene) {
     'use strict';
 
-    var frameState;
-    var camera;
+    var scene;
 
-    beforeEach(function() {
-        frameState = createFrameState();
-        camera = frameState.camera;
+    beforeAll(function() {
+        scene = createScene();
+    });
+
+    afterAll(function() {
+        scene.destroyForSpecs();
     });
 
     it('throws when options.rectangle is undefined', function() {
@@ -49,28 +51,28 @@ defineSuite([
         });
 
         // Inside rectangle, above height
-        camera.position = Cartesian3.fromRadians(0.0, 0.0, 20.0);
-        expect(tile.distanceToCamera(frameState)).toEqualEpsilon(10.0, CesiumMath.EPSILON3);
+        scene.camera.position = Cartesian3.fromRadians(0.0, 0.0, 20.0);
+        expect(tile.distanceToCamera(scene.frameState)).toEqualEpsilon(10.0, CesiumMath.EPSILON3);
 
         // Inside rectangle, below height
-        camera.position = Cartesian3.fromRadians(0.0, 0.0, 5.0);
-        expect(tile.distanceToCamera(frameState)).toEqual(0.0);
+        scene.camera.position = Cartesian3.fromRadians(0.0, 0.0, 5.0);
+        expect(tile.distanceToCamera(scene.frameState)).toEqual(0.0);
 
         // From southwest
-        camera.position = Cartesian3.fromRadians(west - offset, south - offset, 0.0);
+        scene.camera.position = Cartesian3.fromRadians(west - offset, south - offset, 0.0);
         var southwestPosition = Cartesian3.fromRadians(west, south);
-        var expectedDistance = Cartesian3.distance(camera.position, southwestPosition);
-        expect(tile.distanceToCamera(frameState)).toEqualEpsilon(expectedDistance, CesiumMath.EPSILON1);
+        var expectedDistance = Cartesian3.distance(scene.camera.position, southwestPosition);
+        expect(tile.distanceToCamera(scene.frameState)).toEqualEpsilon(expectedDistance, CesiumMath.EPSILON1);
 
         // From northeast
-        camera.position = Cartesian3.fromRadians(east + offset, north + offset, 0.0);
+        scene.camera.position = Cartesian3.fromRadians(east + offset, north + offset, 0.0);
         var northeastPosition = Cartesian3.fromRadians(east, north);
-        expectedDistance = Cartesian3.distance(camera.position, northeastPosition);
-        expect(tile.distanceToCamera(frameState)).toEqualEpsilon(expectedDistance, CesiumMath.EPSILON1);
+        expectedDistance = Cartesian3.distance(scene.camera.position, northeastPosition);
+        expect(tile.distanceToCamera(scene.frameState)).toEqualEpsilon(expectedDistance, CesiumMath.EPSILON1);
     });
 
     it('distanceToCamera in 2D', function() {
-        frameState.mode = SceneMode.SCENE2D;
+        scene.mode = SceneMode.SCENE2D;
 
         var offset = 0.0001;
         var west = -0.001;
@@ -85,19 +87,19 @@ defineSuite([
         });
 
         // Inside rectangle
-        camera.position = Cartesian3.fromRadians(0.0, 0.0, 0.0);
-        expect(tile.distanceToCamera(frameState)).toEqual(Ellipsoid.WGS84.radii.x);
+        scene.camera.position = Cartesian3.fromRadians(0.0, 0.0, 0.0);
+        expect(tile.distanceToCamera(scene.frameState)).toEqualEpsilon(Ellipsoid.WGS84.radii.x, 10.0);
 
         // From southwest
         var southwest3D = new Cartographic(west, south, 0.0);
-        var southwest2D = frameState.mapProjection.project(southwest3D);
+        var southwest2D = scene.frameState.mapProjection.project(southwest3D);
         var position3D = new Cartographic(west - offset, south - offset, 0.0);
-        var position2D = frameState.mapProjection.project(position3D);
+        var position2D = scene.frameState.mapProjection.project(position3D);
         var distance2D = Cartesian2.distance(southwest2D, position2D);
         var height = Ellipsoid.WGS84.radii.x;
         var expectedDistance = Math.sqrt(distance2D * distance2D + height * height);
 
-        camera.position = Cartesian3.fromRadians(position3D.longitude, position3D.latitude);
-        expect(tile.distanceToCamera(frameState)).toEqualEpsilon(expectedDistance, 10.0);
+        scene.camera.position = Cartesian3.fromRadians(position3D.longitude, position3D.latitude);
+        expect(tile.distanceToCamera(scene.frameState)).toEqualEpsilon(expectedDistance, 10.0);
     });
 });